barrier: add working implementation
The following sequence was causing a deadlock earlier -

    pthread_mutex_lock(&bstate.barrier_mutex);
    while (can_enter == 0);

When the current round hasn't finished and a thread leaving it unlocks
barrier_mutex and re-acquires it again, it is stuck in an infinite loop
while holding the lock, thus creating a deadlock. Swapping the order of
above lines fixes the problem.